\name{Bvalue-package}
\alias{Bvalue-package}
\alias{Bvalue}
\docType{package}
\title{ B-Value and Empirical Equivalence Bound
}
\description{
Bvalue package calculates B-value and empirical equivalence bound. B-value is defined as the maximum magnitude of a confidence interval; and the empirical equivalence bound is the minimum B-value at a certain level. A new two-stage procedure for hypothesis testing is proposed, where the first stage is conventional hypothesis testing and the second is an equivalence testing procedure using the introduced empirical equivalence bound.
}
\references{Zhao et al. (2019) \emph{"B-Value and Empirical Equivalence Bound: A New Procedure of Hypothesis Testing"} <arXiv:1912.13084>
}
\author{
Yi Zhao, Indiana University, <zhaoyi1026@gmail.com>

Brian Caffo, Johns Hopkins University, <bcaffo@gmail.com>

Joshua Ewen, Kennedy Krieger Institute and Johns Hopkins University, <ewen@kennedykrieger.org>

Maintainer: Yi Zhao <zhaoyi1026@gmail.com>
}
\keyword{ package }
